wait a bit before force restarting unified

+++ cycle.sh
@@ -2,8 +2,11 @@ if [ -r /afs/cern.ch/user/c/cmst2/Unified/WmAgentScripts/cycle.lock ] ; then
     
     if [ `ps -e -f | grep Uni | grep -c -v grep` == "0" ] ; then
 	echo "There isn't anything running, very suspicious"
-	cat /afs/cern.ch/user/c/cmst2/Unified/WmAgentScripts/cycle.lock /afs/cern.ch/user/c/cmst2/www/unified/logs/last_running | mail -s "[Ops] Emergency On Cycle Lock. Unified isn't running." vlimant@cern.ch,matteoc@fnal.gov,Dmytro.Kovalskyi@cern.ch 
-	rm -f /afs/cern.ch/user/c/cmst2/Unified/WmAgentScripts/cycle.lock
+	sleep 30
+	if [ `ps -e -f | grep Uni | grep -c -v grep` == "0" ] ; then
+	    cat /afs/cern.ch/user/c/cmst2/Unified/WmAgentScripts/cycle.lock /afs/cern.ch/user/c/cmst2/www/unified/logs/last_running | mail -s "[Ops] Emergency On Cycle Lock. Unified isn't running." vlimant@cern.ch,matteoc@fnal.gov,Dmytro.Kovalskyi@cern.ch 
+	    rm -f /afs/cern.ch/user/c/cmst2/Unified/WmAgentScripts/cycle.lock
+	fi
     fi
     echo "cycle is locked"
     exit
